Moved blobDatabase to plugin-core
