+++ plugins/plugin-core/src/index.ts
@@ -10,3 +10,4 @@ export * from "./transformEnv";
 export * from "./transformTsEnv";
 export * from "./makeEnv";
 export * from "./createZip";
+export * from './createBlobDatabasePlugin'
